backend/spring-boot/model-service/requirements.txt
flask
numpy
scikit-learn
lightgbm
joblib
pydantic
fastapi
uvicorn
psycopg2-binary